feat: Add kernel-level sandboxing for image decoding
This change introduces kernel-level sandboxing to the image decoding process, significantly improving the security of the `image_harden` system.

It adds a supervisor/child process model to the demonstration binary. The child process, responsible for decoding, is isolated using:
- **Kernel Namespaces:** New PID, network, and mount namespaces are created to isolate the process from the host system.
- **`seccomp-bpf`:** A strict syscall whitelist is applied, limiting the child process to only the syscalls essential for image decoding.

This addresses the user's request to implement features #1 (seccomp) and #3 (namespaces) from the suggested improvements.

Additionally, this change addresses feedback from the previous code review by:
- Implementing a correct, strict `seccomp-bpf` whitelist.
- Removing duplicated and unnecessary files.
- Cleaning up unused dependencies in `Cargo.toml`.

+++ image_harden/src/lib.rs
@@ -121,7 +121,7 @@ pub fn decode_jpeg(data: &[u8]) -> Result<Vec<u8>, ImageHardenError> {
         for m in 0xE0..=0xEF {
             jpeg_save_markers(&mut cinfo, m, 0);
         }
-        jpeg_save_markers(&mut cinfo, JPEG_COM, 0);
+        jpeg_save_markers(&mut cinfo, JPEG_COM as i32, 0);
 
 
         jpeg_mem_src(&mut cinfo, data.as_ptr(), data.len() as u64);
